@@ -1,20 +1,17 @@
 <template>
-  <v-container>
-    <v-row no-gutters>
-      <v-col cols="12" sm="6">
-        <ul class="footerLinks text-body-1 font-weight-medium">
-          <li><router-link to="/accessibility" exact>Accessibility</router-link></li>
-          <li><router-link to="/privacy" exact>Privacy</router-link></li>
-        </ul>
-      </v-col>
-      <v-col cols="12" sm="6">
-        <p>
-          <CarbonBadge></CarbonBadge>
-        </p>
-      </v-col>
+  <v-footer app color="grey-lighten-5">
+    <v-row justify="center" class="text-xs-center w-100" no-gutters>
+      <v-btn size="large" variant="text" color="info" to="/accessibility">
+        Accessibility
+      </v-btn>
+      <v-btn size="large" variant="text" color="info" to="/privacy">
+        Privacy
+      </v-btn>
+      <CarbonBadge />
     </v-row>
-  </v-container>
+  </v-footer>
 </template>
+
 <script>
 export default {
   data() {
@@ -24,13 +21,4 @@ export default {
 };
 </script>
 
-<style scoped>
-.footerLinks li {
-  display: inline;
-  margin-left: 5px;
-  list-style-type: none;
-}
-.footerLinks li:not(:first-child):before {
-  content: "• ";
-}
-</style>
+<style scoped></style>
