@@ -1,11 +1,8 @@
 <template>
-  <section class="header">
-    <v-container>
-      <h1 class="text-center">{{ title }}</h1>
-      <p class="text-subtitle-1 text-center">{{ subtitle }}</p>
-      <v-divider></v-divider>
-    </v-container>
-  </section>
+  <v-container>
+    <h1 class="text-center text-h2">{{ title }}</h1>
+    <p class="text-subtitle-1 text-center">{{ subtitle }}</p>
+  </v-container>
 </template>
 <script>
 export default {
