@@ -1,11 +1,8 @@
-import Vue from 'vue';
-import Vuex from 'vuex';
+import { createStore } from 'vuex'
 
-import * as jose from 'jose';
+import * as jose from 'jose'
 
-Vue.use(Vuex);
-
-const store = new Vuex.Store({
+export default createStore({
   state: {
     // Logged in user data
     loginKey: null,
@@ -15,48 +12,46 @@ const store = new Vuex.Store({
     loginAdmin: false,
     loginCodes: [],
     // Library service lookups
-    library_services: []
+    libraryServices: []
   },
   mutations: {
-    initialiseStore(state) {
+    initialiseStore (state) {
       if (localStorage.getItem('loginKey')) {
-        state.loginKey = localStorage.getItem('loginKey');
-        const claims = jose.decodeJwt(state.loginKey);
-        state.loginSubject = claims.sub;
-        state.loginIssuer = claims.iss;
-        state.loginExpires = new Date(claims.exp * 1000);
-        state.loginAdmin = claims.admin;
-        state.loginCodes = claims.codes;
+        state.loginKey = localStorage.getItem('loginKey')
+        const claims = jose.decodeJwt(state.loginKey)
+        state.loginSubject = claims.sub
+        state.loginIssuer = claims.iss
+        state.loginExpires = new Date(claims.exp * 1000)
+        state.loginAdmin = claims.admin
+        state.loginCodes = claims.codes
       }
     },
-    setLoginKey(state, key) {
-      localStorage.setItem('loginKey', key);
-      state.loginKey = key;
-      const claims = jose.decodeJwt(state.loginKey);
-      state.loginSubject = claims.sub;
-      state.loginIssuer = claims.iss;
-      state.loginExpires = new Date(claims.exp * 1000);
-      state.loginAdmin = claims.admin;
-      state.loginCodes = claims.codes;
+    setLoginKey (state, key) {
+      localStorage.setItem('loginKey', key)
+      state.loginKey = key
+      const claims = jose.decodeJwt(state.loginKey)
+      state.loginSubject = claims.sub
+      state.loginIssuer = claims.iss
+      state.loginExpires = new Date(claims.exp * 1000)
+      state.loginAdmin = claims.admin
+      state.loginCodes = claims.codes
     },
-    logout(state) {
-      localStorage.removeItem('loginKey');
-      state.loginKey = null;
-      state.loginSubject = null;
-      state.loginIssuer = null;
-      state.loginExpires = null;
-      state.loginAdmin = null;
-      state.loginCodes = [];
+    logout (state) {
+      localStorage.removeItem('loginKey')
+      state.loginKey = null
+      state.loginSubject = null
+      state.loginIssuer = null
+      state.loginExpires = null
+      state.loginAdmin = null
+      state.loginCodes = []
     },
-    setServices(state, library_services) {
-      state.library_services = library_services;
+    setServices (state, libraryServices) {
+      state.libraryServices = libraryServices
     }
   },
   actions: {
-    logout({ commit }) {
-      commit('logout');
+    logout ({ commit }) {
+      commit('logout')
     }
   }
-});
-
-export default store;
+})
