@@ -1,31 +1,21 @@
 <template>
-  <div>
-    <layout-header
-      title="Membership"
-      subtitle="Snapshot counts of library membership by geographic area"
-    />
-    <section>
-      <v-container>
-        <vue-markdown-plus :source="mdText"></vue-markdown-plus>
-        <p class="text-center">
-          <v-btn
-            depressed
-            color="primary"
-            href="https://schema.librarydata.uk/membership"
-            target="_blank"
-          >
-            Membership data schema
-          </v-btn>
-        </p>
-      </v-container>
-    </section>
-  </div>
+  <v-container>
+    <layout-header title="Membership" subtitle="Snapshot counts of library membership by geographic area" />
+    <p class="text-center">
+      <v-btn size="large" variant="tonal" color="info" href="https://schema.librarydata.uk/membership" target="_blank">
+        Explore the membership data schema
+      </v-btn>
+    </p>
+
+    <v-divider inset color="info" class="my-2"></v-divider>
+    <markdown-section :markdownText="mdText" />
+    <v-divider inset color="info" class="my-2"></v-divider>
+
+  </v-container>
 </template>
 
 <script>
-import Header from "../components/layout-header";
 import MarkDownData from "../markdown/membershipdata.md";
-import VueMarkdownPlus from "vue-markdown-plus";
 
 export default {
   data() {
@@ -34,11 +24,7 @@ export default {
     };
   },
   methods: {},
-  components: {
-    "layout-header": Header,
-    VueMarkdownPlus
-  }
+  components: {}
 };
 </script>
-<style scoped>
-</style>
+<style scoped></style>
