@@ -1,283 +1,104 @@
 <template>
-  <div>
-    <layout-header
-      title="Convert library member postcodes"
-      subtitle="Obtain census-based areas for the locations of library members"
-    />
-    <section>
-      <v-container>
-        <vue-markdown-plus :source="mdText"></vue-markdown-plus>
-      </v-container>
-    </section>
-    <section>
-      <v-container>
-        <v-stepper v-model="active_step" flat outlined elevation="0">
-          <v-stepper-header class="elevation-0">
-            <v-stepper-step
-              :complete="active_step > 1"
-              step="1"
-              color="error"
-              editable
-            >
-              Load postcodes
-            </v-stepper-step>
-            <v-divider></v-divider>
-            <v-stepper-step color="secondary" :complete="active_step > 2" step="2">
-              Choose conversion options
-            </v-stepper-step>
-            <v-divider></v-divider>
-            <v-stepper-step color="secondary" :complete="active_step > 3" step="3">
-              Results
-            </v-stepper-step>
-          </v-stepper-header>
+  <v-container>
+    <layout-header title="Postcodes to areas" subtitle="Convert postcode locations to statistical areas" />
 
-          <v-stepper-content step="1">
-            <v-row no-gutters>
-              <v-col cols="12" sm="6">
-                <v-container>
-                  <file-upload v-bind:file="files" v-on:change-files="files = $event" />
-                  <v-btn
-                    color="primary"
-                    depressed
-                    :disabled="files.length === 0"
-                    v-on:click="confirmFile"
-                  >
-                    Next
-                  </v-btn>
-                </v-container>
-              </v-col>
-              <v-col cols="12" sm="6">
-                <v-container>
-                  <v-alert border="right" color="blue" text type="info">
-                    <p>
-                      <b>File tips</b>
-                    </p>
-                    <ol>
-                      <li>
-                        This tool loads CSV files. If your data isn't in CSV format you'll
-                        need to convert it. Try <b>Save as</b> in your software.
-                      </li>
-                      <li>The first row should be column headings</li>
-                      <li>One column should contain UK postcodes</li>
-                    </ol>
-                  </v-alert>
-                </v-container>
-              </v-col>
-            </v-row>
-          </v-stepper-content>
-          <v-stepper-content step="2">
-            <v-row no-gutters>
-              <v-col cols="12" sm="6">
-                <v-container>
-                  <v-select
-                    v-model="postcode_column"
-                    :items="columns"
-                    label="Select column"
-                    outlined
-                  ></v-select>
-                  <v-btn
-                    color="primary"
-                    depressed
-                    v-on:click="confirmOptions"
-                    :disabled="postcode_column === ''"
-                    >Convert
-                  </v-btn>
-                  <v-container>
-                    <v-subheader>My columns include a count</v-subheader>
-                    <v-select
-                      :value="counts_column"
-                      :items="columns.filter((c) => c !== postcode_column)"
-                      :disabled="
-                        postcode_column === '' ||
-                        columns.filter((c) => c !== postcode_column).length === 0
-                      "
-                      label="Count column (optional)"
-                      outlined
-                    ></v-select>
-                  </v-container>
-                </v-container>
-              </v-col>
-              <v-col cols="12" sm="6">
-                <v-container>
-                  <v-alert border="right" color="blue" text type="info">
-                    <p>
-                      <b>Choose your columns</b>
-                    </p>
-                    <p>The column headings in your data should be displayed.</p>
-                    <ol>
-                      <li>Select the correct column for the postcodes</li>
-                      <li>Check guidance below on the optional step</li>
-                      <li>
-                        When ready, select the
-                        <strong>Convert</strong> option to continue
-                      </li>
-                    </ol>
-                    <br />
-                    <p>It may take a few minutes depending on number of postcodes.</p>
-                  </v-alert>
-                  <v-alert border="right" color="orange" text type="warning">
-                    <p>
-                      <b>Optional: counts of people</b>
-                    </p>
-                    <p>
-                      If your data is grouped into counts per postcode, select the column
-                      the count is held in. The tool will then be able to count people per
-                      area for you.
-                    </p>
-                    <p>
-                      If each of your rows relates to an individual person you can ignore
-                      this option.
-                    </p>
-                  </v-alert>
-                </v-container>
-              </v-col>
-            </v-row>
-          </v-stepper-content>
-          <v-stepper-content step="3">
-            <v-row no-gutters>
-              <v-col cols="12" sm="6">
-                <v-container>
-                  <v-subheader>
-                    {{ "Completed in " + getTimeCompleted() + " seconds" }}
-                  </v-subheader>
-                  <v-data-table
-                    dense
-                    disable-filtering
-                    disable-pagination
-                    disable-sort
-                    hide-default-footer
-                    :headers="summary_columns"
-                    :items="summary_data"
-                    item-key="total"
-                  ></v-data-table>
-                  <br />
-                  <v-btn color="primary" depressed v-on:click="downloadConvertedFile">
-                    Save converted file
-                  </v-btn>
-                </v-container>
-              </v-col>
-              <v-col cols="12" sm="6">
-                <v-container>
-                  <v-alert border="right" color="blue" text type="info">
-                    <p>
-                      <b>Saving converted file - what to expect</b>
-                    </p>
-                    <ul>
-                      <li>Valid postcodes are changed to their statistical area</li>
-                      <li>
-                        Postcodes that are no longer valid are changed to
-                        <strong>Terminated</strong>.
-                      </li>
-                      <li>
-                        Postcodes not recognised changed to
-                        <strong>Unknown</strong>.
-                      </li>
-                      <li>
-                        The postcode column header changed to
-                        <strong>Area</strong>.
-                      </li>
-                    </ul>
-                  </v-alert>
-                </v-container>
-              </v-col>
-            </v-row>
+    <v-divider inset color="info" class="my-2"></v-divider>
+    <markdown-section :markdownText="mdText" />
+    <v-divider inset color="info" class="my-2"></v-divider>
 
-            <v-row no-gutters>
-              <v-col cols="12" sm="6">
-                <v-container>
-                  <h4>Library membership data</h4>
-                  <br />
-                  <v-select
-                    v-model="authority"
-                    :items="library_services"
-                    label="Library service name"
-                    outlined
-                  ></v-select>
-                  <v-dialog
-                    ref="dialog"
-                    v-model="modal"
-                    :return-value.sync="extract_date"
-                    persistent
-                    width="290px"
-                  >
-                    <template v-slot:activator="{ on, attrs }">
-                      <v-text-field
-                        v-model="extract_date"
-                        label="Count date"
-                        prepend-inner-icon="mdi-calendar"
-                        readonly
-                        v-bind="attrs"
-                        v-on="on"
-                      ></v-text-field>
-                    </template>
-                    <v-date-picker v-model="extract_date" scrollable>
-                      <v-spacer></v-spacer>
-                      <v-btn text color="primary" @click="modal = false"> Cancel </v-btn>
-                      <v-btn text color="primary" @click="$refs.dialog.save(date)">
-                        OK
-                      </v-btn>
-                    </v-date-picker>
-                  </v-dialog>
-                  <v-btn
-                    text
-                    color="primary"
-                    v-on:click="downloadSchemaFile"
-                    :disabled="authority === '' || extract_date === null"
-                    >Save membership file</v-btn
-                  >
-                </v-container>
-              </v-col>
-              <v-col cols="12" sm="6">
-                <v-alert border="right" color="orange" text type="warning">
-                  <p>
-                    <b>Public library open data</b>
-                  </p>
-                  <p>
-                    Census geographies are required for
-                    <a href="https://schema.librarydata.uk/membership" target="_blank"
-                      >library membership data</a
-                    >.
-                  </p>
-                  <p>This tool has calculated a count of members per statistical area.</p>
-                  <ol>
-                    <li>Select the name of your library service</li>
-                    <li>
-                      Choose a date the postcodes were extracted from your library
-                      management system
-                    </li>
-                    <li>Save the data file</li>
-                  </ol>
-                </v-alert>
-              </v-col>
-            </v-row>
-          </v-stepper-content>
-        </v-stepper>
-      </v-container>
-    </section>
+    <h2 class="text-h5 text-decoration-underline my-3">Convert postcodes</h2>
+
+    <v-sheet color="grey-lighten-5" rounded elevation="0" class="px-5 py-5">
+      <v-alert icon="mdi-numeric-1-circle" class="mb-1" title="Load a file">
+        This tool loads CSV files. If your data isn't in CSV format try <strong>Save as</strong> in your software.
+        The first row should be column headings, one column should contain UK postcodes
+      </v-alert>
+
+      <file-upload v-bind:files="files" v-on:change-files="confirmFile($event)" />
+
+      <v-alert class="mt-8 mb-4" icon="mdi-numeric-2-circle" title="Select file options">
+        The field headings in your data should be displayed below. Choose which one represents the postcodes. If your data
+        includes counts per postcode, also select the field holding the count</v-alert>
+
+      <v-select color="success" v-model="postcodeColumn" :items="columns" label="Select postcode field" variant="outlined"
+        :disabled="columns.length === 0"></v-select>
+
+      <v-select color="success" :value="countsColumn" :items="columns.filter((c) => c !== postcodeColumn)" :disabled="postcodeColumn === '' ||
+        columns.filter((c) => c !== postcodeColumn).length === 0
+        " label="Count field (optional)" variant="outlined"></v-select>
+
+      <v-btn color="info" variant="tonal" size="large" v-on:click="confirmOptions" :disabled="postcodeColumn === ''"
+        append-icon="mdi-chevron-triple-right">
+        Convert
+      </v-btn>
+
+      <v-alert class="mt-8 mb-4" icon="mdi-numeric-3-circle" title="Results">
+        Valid postcodes are changed to their statistical area code. Postcodes that are no longer valid are changed to
+        <strong>Terminated</strong>. Postcodes not recognised are changed to <strong>Unknown</strong>.<br />
+      </v-alert>
+
+      <span v-if="endTime !== null">
+        <p class="mb-2">
+          {{ "Completed in " + getTimeCompleted() + " seconds" }}
+        </p>
+
+        <v-card color="grey-lighten-2" variant="outlined" elevation="0" class="mb-2">
+          <v-data-table no-filter :headers="summaryColumns" :items="summaryData" item-key="total">
+            <template #bottom></template>
+          </v-data-table>
+        </v-card>
+
+        <v-btn color="info" variant="tonal" size="large" v-on:click="downloadConvertedFile"
+          append-icon="mdi-content-save">
+          Save converted file
+        </v-btn>
+      </span>
+
+      <v-alert class="mt-8 mb-4" icon="mdi-numeric-4-circle" title="Library membership data">
+        Census geographies are required to publish <a href="https://schema.librarydata.uk/membership"
+          target="_blank">library membership data</a>. This tool has calculated a count of members per statistical area.
+        Select the name of your library service and choose a date the postcodes were extracted from your library
+        management system
+      </v-alert>
+
+      <span v-if="endTime !== null">
+
+        <service-select v-on:select="authority = $event" />
+
+        <v-text-field class="mt-4" variant="outlined" v-model="extractDate" label="Count date"
+          prepend-inner-icon="mdi-calendar" type="date"></v-text-field>
+
+        <v-btn variant="tonal" color="success" v-on:click="downloadSchemaFile" size="large" append-icon="mdi-content-save"
+          :disabled="authority === '' || extractDate === null">Save membership data</v-btn>
+
+      </span>
+
+    </v-sheet>
+
     <v-dialog v-model="loading" persistent width="300">
       <v-card>
         <v-card-title>
-          <span class="headline">Please wait</span>
+          Please wait
         </v-card-title>
         <v-card-text>
-          <v-progress-linear indeterminate color="primary"></v-progress-linear>
+          <v-progress-linear indeterminate color="info"></v-progress-linear>
         </v-card-text>
       </v-card>
     </v-dialog>
-  </div>
+  </v-container>
 </template>
 
 <script>
 import FileUpload from "../components/file-upload";
-import Header from "../components/layout-header";
 import MarkDownData from "../markdown/membershippostcodetolsoa.md";
-import VueMarkdownPlus from "vue-markdown-plus";
+import ServiceSelect from '../components/service-select'
 
 import moment from "moment";
 
 const config = require("../helpers/config.json");
 
 import { saveAs } from "file-saver";
+
 import * as Papa from "papaparse";
 import * as csvHelper from "../helpers/csv";
 import * as postcodeHelper from "../helpers/postcode";
@@ -285,157 +106,149 @@ import * as postcodeHelper from "../helpers/postcode";
 export default {
   data() {
     return {
-      mdText: MarkDownData,
-      start_time: null,
-      end_time: null,
-      loading: false,
-      loading_message: "",
-      active_step: 1,
+      authority: null,
       columns: [],
+      countsColumn: "",
+      csvData: [],
+      endTime: null,
+      extractDate: null,
       files: [],
-      postcode_column: "",
-      counts_column: "",
-      csv_data: [],
-      lsoas_counted: {},
-      summary_data: [{ total: 0, converted: 0, terminated: 0, unknown: 0 }],
-      summary_columns: [
-        { value: "total", text: "Total" },
-        { value: "converted", text: "Successful" },
-        { value: "terminated", text: "Terminated" },
-        { value: "unknown", text: "Unknown" }
+      libraryServices: config.libraryServices,
+      loading: false,
+      lsoasCounted: {},
+      mdText: MarkDownData,
+      postcodeColumn: "",
+      startTime: null,
+      summaryColumns: [
+        { key: "total", title: "Total" },
+        { key: "converted", title: "Successful" },
+        { key: "terminated", title: "Terminated" },
+        { key: "unknown", title: "Unknown" }
       ],
-      authority: "",
-      extract_date: null,
-      library_services: config.library_services,
-      modal: false
+      summaryData: [{ total: 0, converted: 0, terminated: 0, unknown: 0 }]
     };
   },
   methods: {
-    getTimeCompleted: function () {
-      if (this.start_time && this.end_time) {
-        return Math.round(
-          moment.duration(this.end_time.diff(this.start_time)).asSeconds()
-        );
-      }
-      return "";
-    },
-    confirmFile: async function () {
+    confirmFile: async function (files) {
+      this.files = files;
       let self = this;
       self.loading = true;
       if (self.files.length > 0) {
         const data = await csvHelper.parseFile(self.files[0], false);
         self.columns = data[0];
-        self.csv_data = data;
-        self.active_step = 2;
+        self.csvData = data;
         self.loading = false;
       }
     },
     confirmOptions: function () {
       let self = this;
-      self.start_time = moment();
+      self.startTime = moment();
       self.loading = true;
       // First get all the postcodes
-      self.summary_data[0].total = 0;
-      self.summary_data[0].converted = 0;
-      self.summary_data[0].unknown = 0;
-      self.summary_data[0].terminated = 0;
-      self.lsoas_counted = { Unknown: 0 };
-      const postcode_column_index = self.columns.indexOf(self.postcode_column);
-      const counts_column_index = self.columns.indexOf(this.counts_column);
-      const postcodes = self.csv_data
-        .map((row) => row[postcode_column_index])
-        .filter((p) => p != null && p != self.postcode_column);
+      self.summaryData[0].total = 0;
+      self.summaryData[0].converted = 0;
+      self.summaryData[0].unknown = 0;
+      self.summaryData[0].terminated = 0;
+      self.lsoasCounted = { Unknown: 0 };
+      const postcodeColumnIndex = self.columns.indexOf(self.postcodeColumn);
+      const countsColumnIndex = self.columns.indexOf(this.countsColumn);
+      const postcodes = self.csvData
+        .map((row) => row[postcodeColumnIndex])
+        .filter((p) => p != null && p != self.postcodeColumn);
       postcodeHelper.extractLsoaLookupFromPostcodes(postcodes, (postcode_lookup) => {
-        self.csv_data.forEach((row, idx) => {
-          const postcode = row[postcode_column_index];
+        self.csvData.forEach((row, idx) => {
+          const postcode = row[postcodeColumnIndex];
           const stripped = postcode.replace(/\s/g, "");
-          if (idx === 0 && postcode === self.postcode_column) {
+          if (idx === 0 && postcode === self.postcodeColumn) {
             // This is the header row
-            row[postcode_column_index] = "LSOA";
+            row[postcodeColumnIndex] = "LSOA";
           } else {
             // This is a postcode row
-            this.summary_data[0].total++;
+            this.summaryData[0].total++;
             let lsoa = "";
             if (postcode && postcode !== "" && postcode_lookup[stripped]) {
               lsoa = postcode_lookup[stripped];
               if (lsoa === "Terminated") {
-                row[postcode_column_index] = "Terminated";
-                this.summary_data[0].terminated++;
+                row[postcodeColumnIndex] = "Terminated";
+                this.summaryData[0].terminated++;
               } else if (lsoa === "Unknown") {
-                row[postcode_column_index] = "Unknown";
-                this.summary_data[0].unknown++;
+                row[postcodeColumnIndex] = "Unknown";
+                this.summaryData[0].unknown++;
               } else {
                 // We have a valid lookup
-                row[postcode_column_index] = lsoa;
-                this.summary_data[0].converted++;
+                row[postcodeColumnIndex] = lsoa;
+                this.summaryData[0].converted++;
               }
             } else {
-              row[postcode_column_index] = "Unknown";
-              this.summary_data[0].unknown++;
+              row[postcodeColumnIndex] = "Unknown";
+              this.summaryData[0].unknown++;
               lsoa = "Unknown";
             }
             if (lsoa === "Terminated") lsoa = "Unknown";
             // Ensure our lookup count has the lsoa key value
-            if (!this.lsoas_counted[lsoa]) this.lsoas_counted[lsoa] = 0;
-            if (this.counts_column !== "") {
+            if (!this.lsoasCounted[lsoa]) this.lsoasCounted[lsoa] = 0;
+            if (this.countsColumn !== "") {
               // if count column use that
-              this.lsoas_counted[lsoa] =
-                this.lsoas_counted[lsoa] + parseInt(row[counts_column_index]);
+              this.lsoasCounted[lsoa] =
+                this.lsoasCounted[lsoa] + parseInt(row[countsColumnIndex]);
             } else {
               // else just add one
-              this.lsoas_counted[lsoa]++;
+              this.lsoasCounted[lsoa]++;
             }
           }
         });
         self.active_step = 3;
         this.loading = false;
-        this.end_time = moment();
+        this.endTime = moment();
       });
     },
     downloadConvertedFile: function () {
-      this.downloadFile("converted.csv", this.csv_data);
+      this.downloadFile("converted.csv", this.csvData);
     },
     downloadSchemaFile: async function () {
-      let membership_data = [["Local authority", "Count date", "Area code", "Members"]];
-      const date_string = moment(this.extract_date).format("YYYY-MM-DD");
-      Object.keys(this.lsoas_counted)
+      let membershipData = [["Local authority", "Count date", "Area code", "Members"]];
+      const date_string = moment(this.extractDate).format("YYYY-MM-DD");
+      Object.keys(this.lsoasCounted)
         .filter((k) => k !== "Unknown" && k !== "Terminated")
         .forEach((lsoa) => {
-          let count = this.lsoas_counted[lsoa];
-          membership_data.push([
-            this.authority,
+          let count = this.lsoasCounted[lsoa];
+          membershipData.push([
+            this.authority.name,
             date_string,
             lsoa,
             count > 4 ? count.toString() : "x"
           ]);
         });
-      let count = this.lsoas_counted["Unknown"];
+      let count = this.lsoasCounted["Unknown"];
       // Push unknown and terminated (all as unknown)
-      if (this.lsoas_counted["Unknown"] > 0) {
-        membership_data.push([
-          this.authority,
+      if (this.lsoasCounted["Unknown"] > 0) {
+        membershipData.push([
+          this.authority.name,
           date_string,
           "Unknown",
           count > 4 ? count.toString() : "x"
         ]);
       }
-      this.downloadFile("membership.csv", membership_data);
+      this.downloadFile("membership.csv", membershipData);
     },
     downloadFile: function (filename, data) {
       var blob = new Blob([Papa.unparse(data)], { type: "text/plain;charset=utf-8" });
       saveAs(blob, filename);
+    },
+    getTimeCompleted: function () {
+      if (this.startTime && this.endTime) {
+        return Math.round(
+          moment.duration(this.endTime.diff(this.startTime))
+        );
+      }
+      return "";
     }
   },
   components: {
     "file-upload": FileUpload,
-    "layout-header": Header,
-    VueMarkdownPlus
+    "service-select": ServiceSelect
   }
 };
 </script>
 
-<style scoped>
-tbody tr:hover {
-  background-color: transparent !important;
-}
-</style>
+<style scoped></style>
