@@ -1,31 +1,22 @@
 <template>
-  <div>
-    <layout-header
-      title="Mobile library stops"
-      subtitle="Locations of mobile library stops, and their frequency"
-    />
-    <section>
-      <v-container>
-        <vue-markdown-plus :source="mdText"></vue-markdown-plus>
-        <p class="text-center">
-          <v-btn
-            depressed
-            color="primary"
-            href="https://schema.librarydata.uk/mobile-library-stops"
-            target="_blank"
-          >
-            Mobile libraries data schema
-          </v-btn>
-        </p>
-      </v-container>
-    </section>
-  </div>
+  <v-container>
+    <layout-header title="Mobile library stops" subtitle="Locations of mobile library stops, and their frequency" />
+    <p class="text-center">
+      <v-btn size="large" variant="tonal" color="info" href="https://schema.librarydata.uk/mobile-library-stops"
+        target="_blank">
+        Explore the mobile stops data schema
+      </v-btn>
+    </p>
+
+    <v-divider inset color="info" class="my-2"></v-divider>
+    <markdown-section :markdownText="mdText" />
+    <v-divider inset color="info" class="my-2"></v-divider>
+
+  </v-container>
 </template>
 
 <script>
-import Header from "../components/layout-header";
-import MarkDownData from "../markdown/membershipdata.md";
-import VueMarkdownPlus from "vue-markdown-plus";
+import MarkDownData from "../markdown/mobilelibrarystopsdata.md";
 
 export default {
   data() {
@@ -34,11 +25,6 @@ export default {
     };
   },
   methods: {},
-  components: {
-    "layout-header": Header,
-    VueMarkdownPlus
-  }
+  components: {}
 };
 </script>
-<style scoped>
-</style>
