@@ -1,28 +1,22 @@
 <template>
-  <div>
-    <layout-header title="Physical visits" subtitle="Footfall of people into library premises" />
-    <section>
-      <v-container>
-        <vue-markdown-plus :source="mdText"></vue-markdown-plus>
-        <p class="text-center">
-          <v-btn
-            depressed
-            color="primary"
-            href="https://schema.librarydata.uk/physical-visits"
-            target="_blank"
-          >
-            Physical visits data schema
-          </v-btn>
-        </p>
-      </v-container>
-    </section>
-  </div>
+  <v-container>
+    <layout-header title="Physical visits" subtitle="Counts of people using libraries" />
+    <p class="text-center">
+      <v-btn size="large" variant="tonal" color="info" href="https://schema.librarydata.uk/physical-visits"
+        target="_blank">
+        Explore the visits data schema
+      </v-btn>
+    </p>
+
+    <v-divider inset color="info" class="my-2"></v-divider>
+    <markdown-section :markdownText="mdText" />
+    <v-divider inset color="info" class="my-2"></v-divider>
+
+  </v-container>
 </template>
 
 <script>
-import Header from "../components/layout-header";
 import MarkDownData from "../markdown/physicalvisitsdata.md";
-import VueMarkdownPlus from "vue-markdown-plus";
 
 export default {
   data() {
@@ -31,11 +25,7 @@ export default {
     };
   },
   methods: {},
-  components: {
-    "layout-header": Header,
-    VueMarkdownPlus
-  }
+  components: {}
 };
 </script>
-<style scoped>
-</style>
+<style scoped></style>
