@@ -1,20 +1,11 @@
 <template>
-  <div>
-    <layout-header
-      title="Privacy policy"
-      subtitle="How your data and privacy is protected"
-    />
-    <section>
-      <v-container>
-        <vue-markdown-plus :source="mdText"></vue-markdown-plus>
-      </v-container>
-    </section>
-  </div>
+  <v-container>
+    <layout-header title="Privacy policy" subtitle="How your data and privacy is protected" />
+    <markdown-section :markdownText="mdText" />
+  </v-container>
 </template>
 <script>
-import Header from "../components/layout-header";
 import MarkDownData from "../markdown/privacy.md";
-import VueMarkdownPlus from "vue-markdown-plus";
 
 export default {
   data() {
@@ -23,12 +14,6 @@ export default {
     };
   },
   methods: {},
-  components: { "layout-header": Header, VueMarkdownPlus }
+  components: {}
 };
 </script>
-
-<style scoped>
-.main {
-  margin: 20px;
-}
-</style>
